Turn get_array_length into a unary operation
Previously, the get_array_length instruction was treated as a special
kind of instruction. However, as the length of an array cannot change
once it's allocated, this instruction is actually a pure function and
can thus be turned into a unary operation. This will allow CSE to common
reads of the array length, which can be very important in real-world
programs.