Fix invalid field read of known objects
Previously, the constant folder contained a pattern to fold final field
reads of known objects. This pattern would unconditionally read the
field in question without checking to make sure that the known object
was of the correct type. While it should be impossible to reach code
containing an incorrect field read like this at runtime, it is possible
to generate code that does this that would be unreachable in practice.
If such unreachable code appeared, the constant folder would crash due
to performing an invalid field read. This has been corrected, and a
poison value will be substituted for invalid field reads.