Improve CSE in the presence of intervening blocks
The previous implementation of CSE would always flush all field loads if
the immediate dominator of a block was not the sole predecessor of that
block, i.e. this was not an extended basic block. This was needlessly
restrictive in a lot of cases, as it would cause field loads to not be
commoned across control flow diamonds, even if neither side of the
diamond contained anything that would prevent the load from being
commoned.

To address this, CSE has been improved to use the union of the kill sets
of all "intervening blocks", i.e. all blocks that could run between the
end of the immediate dominator and the start of the current block. This
allows much more aggressive commoning of field loads by CSE.