Fix phi nodes being placed on entry block
Under certain conditions, it was possible for block merging to merge the
first and second blocks of a function when the second block contained
phi nodes. This would result in phi nodes on the entry block, which
caused chaos during inlining and codegen. To correct this, the block
merger has been adjusted to avoid placing phi nodes on the entry block
and the validator has been updated to check that the entry block does
not have any phi nodes.