Remove the get_param instruction
Previously, the only way to get the value of an argument in MIL was to
execute a get_param instruction. This is completely unnecessary, as this
value will never change during execution of the function. Instead, the
function now contains a list of registers for the parameters to be
placed in, which can be referenced in the same way as any other
register.