Print register targets before instruction names
Previously, the register target of an instruction was printed as if it
were the first operand to the instruction, similar to Intel x86 assembly
format. This made it quite difficult to scan through printed MIL
instructions to find the instruction that was setting a particular
register. To correct this, the target register is now printed before the
instructon name, as an assignment expression. This is more similar to
how LLVM IR is printed and makes the printed instructions much more
readable.